(inprog) Fix viewer issues
Viewer now utilises a grid and has a newly written in-page navigation
system. The actual rendered object also constrains to the browser as
best it can, and I made an attempt to fix the previous scaling issues
that came from overuse of viewport sizing classes.

The internal navigation follows the same semantics as react-router and
allows for in-page navigation based on a query string tha tcan be set
via history methods. It's not as sophisticated as the other react-router
router implementations, but it gets the job done.