Media viewer (#2)
Add front-end and back-end bindings for viewing media. An API route has
been created to fetch metadata necessary to render an upload. A new
in-page navigation system was also written to handle tab-like views on
the viewer page.

The internal navigation follows the same semantics as react-router and
allows for in-page navigation based on a query string that can be set
via history methods. It's not as sophisticated as the other react-router
router implementations, but it gets the job done.

Tailwind's config was modified to remove the image/video sizing resets
from their preflight as they were preventing the render of unconstrained
media.