(inprog) Profiles
Back-end bindings.

This commit brings a large refactor to the database bindings - pretty
much every fetcher method now allows you to reuse a handle. Since
there's no limit to the amount of uploads someone can own, this was the
best way to not exhause our connection pool when extending 1k+ uploads
in a single batch.