Media Processor
This commit includes the v1 Media Processor code.
Messaging is implemented using ZeroMQ: Consumers REQ a job from the
producer, who REPs with a BSON-serialized object of data containing
everything needed to process. When the consumer is done, it PUSHes the
result to a consolidator which updates the database. MediaProcessor
relies on FFMPEG and the README has been updated to mention this.

A justification for some implementation choices:

I wrote the media processing queue to use ZeroMQ instead of supplying
data directly to a class so that I can distribute processors over the
network. There are however limitations to that with the current app
infrastructure given that files are stored on the local disk. When we
move over to a distributed fs where workers can gain access easier
across a network then there's a chance that the processors will be
distributed in clusters. Writing the code to be distributable now
solves the complexity later. There's also a chance that the processor
will be extracted from the codebaes and rewritten in another language
for distribution, e.g. nodejs which handles event-driven workloads
excellently.

BSON was chosen as the package de/serializer because of its adoption in
MongoDB. I explored Procbuffer and Avro, but MongoDB was better for our
use case - it allows us to write our own serializers and doesn't require
code generation or sending schemas over the wire with each packet. It
also has adoption in most other languages and will allow for better
portability across different languages.