Comments
Comments are markdown-enabled and include link resolution in the form of
`anchorme`. Links in comments redirect to the `/leaving` endpoint which
gets confirmation from the user before allowing them forward. Markdown
is statically generated and stored in the database, links are
dynamically generated to handle edge cases in the future.

Both the back-end and front-end respect the COMMENTS_RESTRICTED flag on
accounts and display warnings as appropriate.

(inprog) Comments back-end bindings