Basic relational tagging infrastructure added
Implement a basic setup for relational-based tagging. Add a TagManager
as the application's "source of truth" for the DB's tag state. The
TagManager uses a RadixTree to facilitate wildcard searches and acts as
a tag cache go-between.

Some basic CLI commands have been implemented to interface with the
TagManager and update tag states on uploads.