Tagging (#7)
Tagging has been implemented using a hybrid approach: tags and metadata
are stored in postgres and tagged relations are stored in both postgres
and elasticsearch. Relations are stored in both for ease of aggregation.

Tags are implemented using a TagManager which acts as a source of truth
on the database state while maintaining an internal memory cache using
a ConcurrentRadixTree. This allows us to do multiple different types of
tag searches without hitting the database, and the tree allows us to
short-circuit prefixed searches. This gives us a minor efficiency boost
and still maintains state.

Tags come in two flavors: userland and system. System tags are
auto-tagged by the MediaProcessor and include items such as 
"filesize:massive" and "meta:has_audio". This allows users to know for a
fact that what they're requesting is something that was auto tagged as
opposed to something user tagged. Userland tags are not allowed to
mimic system tags but that is their only constraint.

Search is implemented in Elastic using a custom TagTokenizer that takes
user input and maps it to a consumable BoolQuery. Searching can only be
ordered by upload ID at this point in time, however it is planned to
expand metadata on the upload object to include items such as score,
votes, video length, and anything else a user may want to order by.