Replace FontAwesome with MaterialDesignIcons
FontAwesome is a great resource but having used the pro icons in pretty
much every personal project I've worked on, having to go back down to
the free tier because FA hasn't figured out how to allow pro icons in
open source yet means there are a lot of places where I either need a
specific icon that I can't access, or I need a specific feature that I
can't access. The final nail in the coffin was having to remove the
React component and go back to using the font-based icon format which
introduces issues when network drops.

As a result, I've made the decision to move over to MaterialDesignIcons
for this project. It exposes a React component, it allows me to switch
back to using SVG icons, and it has a massive library of icons that
outshines the FontAwesome free tier in a way that's simply hard to
ignore. If FontAwesome figures out how to support Pro tier in open
source like tailwindui has then I'll switch back. But until then, MDI
will be the main icon driver for the project.